+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - when running `pineappl convolute ... -s 1` the scale-variation columns are no
   longer shown. The output would be zero, but this doesn't make sense. All
   other values of `-s` are unaffected.
+- added a further optimization to `Grid::optimize` that strips empty orders
 
 ### Fixed