+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.1] - 01/03/2022
+
 ### Added
 
 - added the `--bins` option to the CLI subcommand `sum`, which allows to sum
@@ -214,7 +216,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - first release
 
-[Unreleased]: https://github.com/N3PDF/pineappl/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/N3PDF/pineappl/compare/v0.5.1...HEAD
+[0.5.1]: https://github.com/N3PDF/pineappl/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/N3PDF/pineappl/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/N3PDF/pineappl/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/N3PDF/pineappl/compare/v0.3.0...v0.4.0